move init_db() into database module
