+++ reposcan/database/database_handler.py
@@ -1,6 +1,7 @@
 """
 Module containing database handler class.
 """
+import os
 import psycopg2
 
 
@@ -33,3 +34,11 @@ class DatabaseHandler:
         """Rollback any pending transaction."""
         if cls.connection is not None:
             cls.connection.rollback()
+
+def init_db():
+    """Setup DB connection parameters"""
+    DatabaseHandler.db_name = os.getenv('POSTGRESQL_DATABASE', "vmaas")
+    DatabaseHandler.db_user = os.getenv('POSTGRESQL_USER', "vmaas_writer")
+    DatabaseHandler.db_pass = os.getenv('POSTGRESQL_PASSWORD', "vmaas_writer_passwd")
+    DatabaseHandler.db_host = os.getenv('POSTGRESQL_HOST', "database")
+    DatabaseHandler.db_port = os.getenv('POSTGRESQL_PORT', 5432)
